refactor(User): move to new folder
