+++ p-week-01/TodoApp/Models/User/UserModel.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace TodoApp.Models
+namespace TodoApp.Models.User
 {
     public class UserModel
     {
