+++ p-week-01/TodoApp/Repositories/ApplicationContext.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using TodoApp.Models;
+using TodoApp.Models.User;
 
 namespace TodoApp.Repositories
 {
