+++ p-week-01/TodoApp/Services/AuthService.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TodoApp.Models;
+using TodoApp.Models.User;
 
 namespace TodoApp.Services
 {
