+++ p-week-01/TodoApp/Services/IUserService.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using TodoApp.Models;
+using TodoApp.Models.User;
 
 namespace TodoApp.Services
 {
