+++ p-week-01/TodoApp/Services/UserService.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using TodoApp.Models;
+using TodoApp.Models.User;
 using TodoApp.Repositories;
 
 namespace TodoApp.Services
